refactor: add a shared base for local executor proxy and runtime executor proxy
